+++ packages/image_selector/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+Make it possible to read what the selected image is, in a form in a different app.
+Exposes `onImageSelected` as a `ValueChanged<SelectedImage>`.
+
 ## 0.0.1
 
 Adding ImageSelectorWidget